backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PDF processing libraries
PyPDF2>=3.0.0
pdfminer.six>=20221105
pdfplumber>=0.10.0
pikepdf>=8.0.0
PyMuPDF>=1.23.0  # fitz module

# Image processing
Pillow>=10.0.0

# PDF generation and manipulation
reportlab>=4.0.0

# Table extraction (optional but useful)
camelot-py[cv]>=0.11.0

# Additional utilities
python-magic>=0.4.27  # For better file type detection
pydantic>=2.0.0  # For data validation
python-dateutil>=2.8.0  # For date handling

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI applications